Extract code and create service
